updating arranged version of app.py
