--- 2023-11-27T22-38_export.csv:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=http://172.17.48.160:8501/
-HostUrl=http://172.17.48.160:8501/
